chore: unify kubeutils
