+++ operator/pkg/upgrade/util.go
@@ -1,16 +1,17 @@
-package k8sutil
+package upgrade
 
 import (
 	"context"
 	"fmt"
+	"strings"
+
 	"github.com/replicatedhq/embedded-cluster/kinds/types"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"strings"
 )
 
-// K0sVersionFromMetadata takes versions like v1.30.5+k0s.0 and returns v1.30.5+k0s to match the kubeletVersion in a cluster
-func K0sVersionFromMetadata(meta *types.ReleaseMetadata) string {
+// k0sVersionFromMetadata takes versions like v1.30.5+k0s.0 and returns v1.30.5+k0s to match the kubeletVersion in a cluster
+func k0sVersionFromMetadata(meta *types.ReleaseMetadata) string {
 	if meta == nil || meta.Versions == nil {
 		return ""
 	}
@@ -22,8 +23,8 @@ func K0sVersionFromMetadata(meta *types.ReleaseMetadata) string {
 	return desiredVersion
 }
 
-// ClusterNodesMatchVersion returns true if all nodes in the cluster have kubeletVersion matching the provided version.
-func ClusterNodesMatchVersion(ctx context.Context, cli client.Client, version string) (bool, error) {
+// clusterNodesMatchVersion returns true if all nodes in the cluster have kubeletVersion matching the provided version.
+func clusterNodesMatchVersion(ctx context.Context, cli client.Client, version string) (bool, error) {
 	var nodes corev1.NodeList
 	if err := cli.List(ctx, &nodes); err != nil {
 		return false, fmt.Errorf("list nodes: %w", err)
